requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
python-dotenv
pytest
httpx
fastapi-users
pymongo
scikit-learn
numpy
pandas
reportlab
fpdf2
matplotlib
seaborn
pillow
seaborn